First try of the new build method to preserve the downloads off the servers, but still start from scratch.
build scripts are parsed for the git or svn urls and they are downloaded to a folder. when they try to run again with the stuff in the folder, SVN updates and git fails. The build scripts for GIT call git pull. the urls are pared because they are variables. they can still be run independantly.

need testing.

need to find a way to automatically remove extranious packages, that are installed from a previous attempt. Need to find a way to remove them when they are no longer specified.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@173 b52b6941-3400-464c-9f42-43200397181c