Dropping the use of ccache, and replacing with checkinstall. ccache did not reduce the amount of time for QT and kdelibs drastically enough to use 12 gigabytes of data.
checkinstall creates debs out of built source
since not all packages need to be rebuilt every ISO build (like QT and KDELIBS) some can be saved and just simply reinstalled untill later.

This is only the first step. the next step is to heavily modify the build scripts to handle the creation of the deb files, and conditions for when to rebuild, and when to not rebuild...

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@800 b52b6941-3400-464c-9f42-43200397181c