attempt to create a front end to loading mplayer for the user
the select-file script gets passed the command that is used to open the file, and it pops up a zenity dialog for the user to select the file, and then runs the command passed to the script, and it passes the file name at the end.
this is usable for not just mplayer, but for the weston viewers, and possibly gstreamer.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@928 b52b6941-3400-464c-9f42-43200397181c