fix some weird quoting in the command for the autologin user
su wants to be run on a terminal. When I start a job in the background, it's not on a terminal. That's why the unit tests worked, but it fails on the real deal, because the loginmanager starts the authenticate user another job, so that the loop doesn't stop, and the fifo can be read. The script command creates a tty for a specified command.

Start the authentication for expect with the script command, so that it has a tty.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@1194 b52b6941-3400-464c-9f42-43200397181c