script doesn't seem to like to work with return values. or screen.
so that su doesn't keep running fork weston, and get the last pid called in the session caller.

Reserve tty 63 for running su. force it to run.  I now made it harder for su conflicts on tty63, by making it fork weston, and return the PID of WESTON to the loginmanager

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@1252 b52b6941-3400-464c-9f42-43200397181c