Retry piping the password into the expect shell, so that it's not stored in an environemnnt variable. (It's stored in an _expect_c variable, but that's not visible to /proc/<pid>/envir)
Replace the su with sudo. sudo isn't pesky about TTYs, and it does not eat pipes,  or return codes, so I can get TTY 63 back

Expect _does_ create the ptys for the su that it calls, it was the su that runs it as deamon that was complaining. 

I run expect as daemon so that su still prompts for the password, and not automatically run



git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@1292 b52b6941-3400-464c-9f42-43200397181c