unset the NewSeatNumber each time
Fix the zenity command so that the title argument for zenity gets passed correctly

Fix the display of the devices in the seat window. I was using the array wrong, and they weren't showing up.

Some devices have the device type in the device string such as "Company 102 key Keyboard", some don't. address this by putting the device type first in the list dialog and config dialog

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@1428 b52b6941-3400-464c-9f42-43200397181c