make the waylandloginmanager open 1 diagnostic root terminal on loginmanagerdisplay if the option wlmdebuginsecure is passed to the kernel command line. If this option is passed, it is determined before the loop begins, so it can only be started once.
As a root teminal can be achived by adding init=/bin/bash to the kernel command line with GRUB, this does not change the effective security in my opinion.
If this option is passed, it logs as it starts the single root terminal (which is only useful if syslog is configured to copy logs to a syslog server)


This can be secured by password protecting the boot manager, which requires physical access. The only way to exploit this during runtime is to bind mount a phony file over /proc/cmdline, and restart the waylandloginmanager service, which would require root access



git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@1678 b52b6941-3400-464c-9f42-43200397181c