Attempt to bring back the package installer from commit 1577 that I dropped in commit 1582, I dropped it because of reliablility issues. Now that I have fail tracking I can add it. And since the fail tracking is subtractive in the sense that it only removes a package from the fail list when successful it's actually more dependable then what I had before.
All new packages are added to the fail list. If it fails to install (or the script is aborted) t stays on the fail list. If it completes, it is removed. On the next run, all new packages diffed out from the new and old versions of installs.txt are installed, and the packages that are in the faillist are appended to be run again.




git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@1717 b52b6941-3400-464c-9f42-43200397181c