Use Debian as the Tier 1 distribution, for Tier 1 packages. Casper and Lupin have been made to be included in Debian, and Calamares is now the installer, as debian has Ubiquity, and it seems hard to find debian-installer For Debian the INSTALLS.txt have been made for conditional install, so that if ever needed, it could be easy to jump back to Ubuntu. Also it supports conditionals based on package architecture as debian doesn't have a linux-generic type package The new Debian CDs now boot to waylandloginmanager, and weston starts.
There are a few things to resolve. packages need to be built from scratch to get the correct dependancies again.
Plymouth settings and the theme doesn't apply, for some reason the rbos-rbos package fails to build (as of now)
And it's not compatible with Ubuntu USB Creator, however mismatched Ubuntu versions have been known to cause issues with that thing. will test with unetbootin

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@3790 b52b6941-3400-464c-9f42-43200397181c