build_core: attempt to improve security by using a restricted shell. it is run with a custom PATH set to a temporary folder with root only permissions that is created, and cleared, and where the revisions file is copied to, as the revisions file uses no external non bash commands. No other executables, even awk, vi, tee, or bash, or sh can be called by this restricted shell, which should further protect against a non valid revisions file from 'escaping' the restricted shell
A valid revisions file has functions OverrideRevisions_($SOURCENAME) for each defined package. When the function is run, it sets the SOURCEREVISION variable, and SUBMODULEREVISIONS variable

The restricted shell executes the revisions file as a 'source' script to declare all the functions, and then runs the function for the current package being built. The SOURCEREVISION and SUBMODULEREVISIONS variables are then exported to STDOUT, by bash, in a format that uses the bash 'declare'  command for preserving arrays.
As the restricted shell can't redirect STDOUT, the main script redirects the output to a new file, and sets it as executable.

The generated file is then sourced in by the main script

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@3912 b52b6941-3400-464c-9f42-43200397181c