Support using scalable tmpfs based on how much RAM is avalible to save wearing out SSDs as much as possible from temporary writes Based on how much RAM is free, an amount is set aside, and paticular features are redirected with bind mounts to use the RAM disk, based on the size of the RAM disk
Right now, the sizes need to be tuned, it's assuming that these need much more then they actually will take up


compile_all: cleanup the last package. Do it AFTER, so that if  one package build needs to be debugged directly, the pause statement can be placed easier, on the line, without having it deleted.
This part is still pending the overlayfs fix, for the srcbuild overlay to be enabled

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@4471 b52b6941-3400-464c-9f42-43200397181c