main builder: Fix the logging to actually work, call the namespaced processes with python in their own PTY so that it's line buffered, and not fully buffered, then capture the output into tee. The last "fix" caused it to not be logged.
The workaroud to call the ENTIRE thing in python is not needed, as the namespaced processes are called in it individually now

Detect a suitable python binary to use, not all systems have 'python', just ones with version numbers

The variable unset loop was causing problems with Python and CTRL+C oddly enough. not the variables that it was unsetting. It was something with

----------------------------------------------
while read var
do
  :
done < <(echo random| pipe)
----------------------------------------------
would then cause python when calling the pty'd process to ignore CTRL+C

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@4950 b52b6941-3400-464c-9f42-43200397181c