waylandloginmanager: when a session fails to start, the session number gets recycled. When that happens, it tried to use the same systemd service to run the user session in. Restart the existing service status, before trying to start it.
Turns out this was a problem that was within waylandloginmanager all along, and not a problem caused by the new parsing.

Enhance the Zenity command for the session chooser, the headers are visible in that one, so capitalize the collumn names. No need for a output seperator, as it only outputs one feild now

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@4962 b52b6941-3400-464c-9f42-43200397181c