build_vars: LIBRARY_PATH and PKG_CONFIG_ALLOW_SYSTEM_CFLAGS probably do not need to be set. Remove them This was causing problems building QtWebEngine 5.12. (Mostly the PKG_CONFIG_ALLOW_SYSTEM_CFLAGS)
Removing PKG_CONFIG_ALLOW_SYSTEM_CFLAGS from the environment allowed QtWebEngine to work
Trying to trace down WHY this variable was set, and it turns out that it has been being set in the scripts that eventually became this project, before they where even under my version control, so has been set for building wayland since commit 1. 

It came from following following the Wayland web building page in late 2011. 

Looks like the variable was added to the how to
https://cgit.freedesktop.org/wayland/wayland-web/commit/?id=f9e9b28cf244e83e8169a34ba848c4ea7fc1d2aa (2011-03-22)
and removed from wayland-web
https://cgit.freedesktop.org/wayland/wayland-web/commit/building.html?id=7c47b695e815ff40ecbb96f58536f7957288fb72 (2012-03-22)
which was then described as not needed. It likely was a workaround that was needed in the early days of Wayland.

Because of this, the variable was set since the beginning of the project
https://sourceforge.net/p/rebeccablackos/code/1/tree//rebeccablacklinux_files/usr/bin/compile/B000wayland

 and even in my scripts that where from BEFORE the project was actually started
https://lists.freedesktop.org/archives/wayland-devel/2011-December/001746.html


A rebuild of everything will be needed to make sure the removal of these variables break nothing.


git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@5869 b52b6941-3400-464c-9f42-43200397181c