weston wat r u doing weston STAHP
In seriousness, this only impacts the tiny content-protection test that comes with Weston. Nothing else should be using this silly protocol in Weston, as it got NAK'ed by the wlroots guys. But it's the spirit that counts. This should in theory allow the content-protection application to set whatever it needs to set for its surfaces, regaurdless of config. But not actually enforce it. And to make the protocol always on so that it can report to the client that protection is enabled... ...when it's not.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@6153 b52b6941-3400-464c-9f42-43200397181c