waylandloginmanager: The flickerfree loginmanagerdisplay (from splash to greeter) apparently was causing some timing issues on the second seat.
the waylandloginmanager is timed to where it starts after the systemd-user-sessions service (as the weston greeters need it) and hold off the service that removes the splash screens, until the service declares itself 'ready'
The waylandloginmanager (if the startup is sucessful) sends this signal after the first loginmanagerdisplay is started up.

This was causing an apparent timing issue where the plymouth service must do something to the /dev/dri/cardX as the second seat, weston will prepetually spam in the loginmanagerdisplay log that weston atomic has "permisson denied", and keep showing the remnants of the splash screen.

This fixes it, by running the same command as the plymouth-quit-wait service, to wait for Plymouth to terminate, before going on setting up the loginmanagerdisplays on other seats. (if any)


git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@6279 b52b6941-3400-464c-9f42-43200397181c