external builders: Don't bind in /sys . /sys isn't really needed during dpkg setup or compilation. (/proc and the mininal /dev are) Hide /proc/modules as well. the console-setup package calls lsmod, which in turn reads /proc/modules .
It seems the console-setup dpkg package setup script filters if the "speakup" module loaded, and configures the console differently
So this works around it so that the configuration (but not down to hashes) will be more likely to be reproducible on any system built

Proving that not binding in /sys is more than just theory even if it is one package.

Testing rebuilding, and nothing failed due to this change.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@6439 b52b6941-3400-464c-9f42-43200397181c