remastersys: switch the ISO bootloader to GRUB instead of ISOLINUX.
On flawed hardware, (such as the emulated vmware card on QEMU_ the firmware framebuffer (efifb/vga/vesa) is a viable fallback.
the vmware card on QEMU offers no mode setting support.

ISOLINUX had the odd flaw of clearing the VGA mode, causing there to be no fallback framebuffer
Turns out the framebuffer is handled by the BOOTLOADER, and not the OS. 
*This* explains the then odd issue in early versions of VirtualBox there was no /dev/fb0 as a live cd, but there would be one when installed.
It was not something odd in casper, it was the ISOLINUX bootloader, where GRUB was handling it better.

Passing vga=current did not seem to work correctly for ISOLINUX, even though it is documented which is odd, so the transition to GRUB was made. 

The ISOs now boot with preserved framebuffers. And now, the simplekms driver that has been poor-man's cherrypicked relies on the frame buffer
These ISOs now better ensure Kernel Mode setting with the default boot option

With the framebuffer being preserved, GRUB also needs far fewer boot options, instead of having a whole bunch hardcoded for different screen resolutions.
These numerous options were confusing users, despite efforts to clarify the help text.

- It still boots in an EFI VM (QEMU with TianoCore). EFI was previously already using GRUB, but now it's a shared config
- It still boots as an ISO in a QEMU VM
- It still boots as a raw hda device in a QEMU VM. (This is to test ISOHYBRID). It works in MBR, not by the EFI VM, but this is not a regression. Old ISOs would not get loaded as a raw device by TianoCore, and neither do Ubuntu ISOs


To ensure maximim compatibility, it seems Ubuntu also switched from ISOLINUX to GRUB in their newer ISOs. 
https://bazaar.launchpad.net/~ubuntu-cdimage/debian-cd/ubuntu/view/head:/tools/boot/groovy/boot-amd64 was used to generate much of the xorriso command
and why a patched version of xorriso was needed.



The ISOLINUX configuration and structure is still maintained as the ISO will be compatible with older UNETBOOTIN still. (although it won't have the same features that GRUB does)
So that is why the isolinux.cfg with the numerous options is still preserved. 

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@6766 b52b6941-3400-464c-9f42-43200397181c