waylandloginmanager: Fix off-by-one error in GetNextTTY to start on TTY7 instead of TTY8
Since the waylandloginmanager was first written in 2013, the greeter would start on TTY8, instead of the more traditional TTY7
In earlier versions, it would start from the first TTY, and end up on TTY8. The cause was something locking TTY7
what was locking it was not investigated before it was hardcoded in a way that would make it use TTY8. 
It was likely upstart, or now that vt.handoff=7 is no longer passed.

waylandloginmanager: Don't reserve TTY63 anymore in GetNextTTY. TTY63 was used for a really old workaround that hasn't been used in a while.
While TTY63 IS used for the text mode fail dialogs if the greeter doesn't start, the greeter has been already working if it started 56 sessions
(63 TTYs minus 6 gettys, minus 1 for the greeter)



git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@6831 b52b6941-3400-464c-9f42-43200397181c