Start the new design for vTTYs
This is similar to the first design. The first design used tmux as client/server.
tmux ended up being too much for being a login handler. Too many configuration options needed to be set to make it 'kiosk ready'
vte and cage for the current design were reluctantly made to be run as root

Now instead, socat is used. socat has the benefit of granting a user access to the client socket, so the pty can be forwareded.
The socat cleint that connects doesn't have to run as root, which means cage and vte don't have to run as root
socat doesn't have tiling, or ways to start new shells, so it's a better fit for running /bin/login

The login sessions run on seat-vtty, so that the fact that XDG_VTNR has to be 0 won't stop pam from working.
nmtui for example is able to handle connections from these sessions using its PAM actions

This hasn't been cut over yet

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@6840 b52b6941-3400-464c-9f42-43200397181c