Overhaul rbos-add-framebuffer to properly handle grub graphics, rename it to rbos-configure-simplegraphics
It will likely not be as relevant even on hardware that requires simple graphics. Testing on real hardware and grub can properly detect a 1366x768 laptop. Probably would be more relevant on VMs though, as it tends to favor 640x480 by default.

Grub could be forced to use a higher resolution on the ISO like it was before, but that broke the proper detection of 1366x768

Add hwinfo. rbos-configure-simplegraphics uses it, while it is not fuly reliable, it is more reliable on the VM hardware where it is more likely to be needed

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7054 b52b6941-3400-464c-9f42-43200397181c