Make the boot more reliable, don't change into graphical mode untill while booting the selection
While this makes the live cd boot not as nice looking, on the QEMU cirrus card, while in the default resolution, the Grub display jumps around and is cropped odd, as the menu is scrolled.
The glitches don't happen while it's booted, but it inhibits the usability of the Grub menu off the live cd mode.
There's no telling if there is actual real hardware where this problem happens
This only happens on the cirrus card in QEMU, while in the autodetected resolution. Forcing it to 1024x768 it does not happen, but then that would force all hardware that needs simplekms/simpledrm to use 1024x768 by default, when say it can support 1366x768 automatically just fine otherwise

Ubuntu's grub is configured to use gfxmode=console, so this must be why.

As a side effect though, this makes setting a custom resolution easier.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7073 b52b6941-3400-464c-9f42-43200397181c