waylandloginmanager: Don't export WAYLAND_DISPLAY to the display server.
kwin, when the --drm option is not able to be passed, it uses the WAYLAND_DISPLAY variable to try to run a nested session.

This has been done since this part of waylandloginmanager was in westonlaunchcaller since commit 2759 . It was likely intended to try to force paticular Wayland display servers to use the desired display that is determined by waylandloginmanager, for display servers that support it. (Similar to how the Display Server that weston uses is set by the --socket option)

It very likely was "swc" that this was for. No existing sessions use the existing WAYLAND_DISPLAY variable to set the socket name for the server.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7144 b52b6941-3400-464c-9f42-43200397181c