Redesign the fullscreen terminal session to use similar design around vTTYs
UvTTYs (User Virtual TTYs) use the socat client-server only instead of a login terminal, it runs the users shell and the interacting terminal, both as the user.
Contrast with vTTYs for tty1-tty6, where one runs a getty as root, and the other runs a display server as a non-root service user.

The Display Server for a UvTTY can be suspended with 
    utty-launchsuspend weston




git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7310 b52b6941-3400-464c-9f42-43200397181c