Try to make the Grub config comment more clear.
While not as confusing as the old ISOLINUX options with multiple framebuffer size options, users that open the edit mode might see it, and think they need to configure a resolution. This is not needed on all hardware. Only hardware that uses SimpleDRM. This should hopefully be clear in simple terms, that don't wrap to two lines on small resolutions

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7426 b52b6941-3400-464c-9f42-43200397181c