UvTTY: Don't unset WAYLAND_DISPLAY
While running Wayland clients under cage when the main terminal emulator is running (currently foot) is not recommended, Wayland clients default to assuming WAYLAND_DISPLAY is "wayland-0" which is also a valid display server name for cage if it's the first display server the user runs. This could lead to an inconsistent case where users can start Wayland clients under the first instance they start, and then being unable to start them on the second one. 

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7427 b52b6941-3400-464c-9f42-43200397181c