Attempt to fix extremely rare race condition where sometimes the vtty display server appears after the waylandloginmanager greeter
Replicating it is very difficult, but it happened on a VM instance when the host was under heavy load. The theory is that tty1 could have been active, and the active tty inhibitor skipped, and started Cage, and Cage could have started slower, waylandloginmanager could have started soon after, and then Cage finished starting and took control. This is just a theory though. The time this happened was not captured by systemd-analyze. With this change however, it was looked at with systemd-analyze, and it seems that there is a benefit that vtty-frontend@tty1.service starts on-demand now, the first time tty1 is manually active. With this, this is very likely to fix the race condition

This has been tested, and the vTTY starts still if the waylandloginmanager service is disabled, and if it fails to start

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7428 b52b6941-3400-464c-9f42-43200397181c