uvtty: Speed up the detection of the session ending, by writing into the FIFO
The session runs in an instance of socat called by the backend, but the backend runs the socat instance in a background process
The backend is responsible for listening to the frontend, that writes the current size of the terminal when it gets the SIGWINCH signal, (when the terminal is resized), it writes to the FIFO
The backend also traps for the exit of the socat process

The FIFO only loops every 5 seconds since resizes of a fullscreen window should be pretty rare, except for say perhaps changing a monitor for less CPU impact
Bash's trap handler for watching the backend exit doesn't handle until after the 5 second timer is up.

Work around this by writing a line into the FIFO to force Bash on the backend to exit, as the PID for the session waits for uvtty-backend to determine if the session is running. The quicker it exits after socat exits, the quicker waylandloginmanager acts to switch back to the loginmanagerdisplay

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7463 b52b6941-3400-464c-9f42-43200397181c