vtty/uvtty: wait a bit after SIG_WINCH before looking up the size, and don't log errors in the frontend.
There still is a rare race condition however it seems to be in foot itself. It was observable graphically before the default background color was changed
It seems that when the first PtyResizeEvent is run from the trapped SIG_WINCH foot is still the initial size, and stty looks up the resolution too quick.


git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7536 b52b6941-3400-464c-9f42-43200397181c