build_core: adopt for the first element ACLOCAL_PATH to be /tmp/aclocal instead of /opt/share/aclocal in ACLOCAL_FLAGS
Since commit 1, ACLOCAL_FLAGS was set to the secondary $PREFIX/share/aclocal, back when it was still /usr/local
This was a mistake. It seems that autoreconf tries to copy all items to the first folder element in ACLOCAL_PATH which, not only is configure not running as root, but it is not captured by checkinstall. (An issue with aclocal was actually a large roadblock before this project was officially created)

Vala depends on m4 files that no other packages have needed so far. This is why this appeared now

This is why this is appearing now
currently, all the m4 files are correctly installed through checkinstall. 

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7539 b52b6941-3400-464c-9f42-43200397181c