waylandloginmanager: Replace the old authentication validation that used exect around su with one that uses perl and PAM. It's a better way to handle it, and there should be less odities with having to escape characters.
Issues with using expect:
it was found (and worked around) that backslashes in the password caused problems with expect in early 2018. 
The test started /bin/true as the logging in user. And then later on it would have to start the actual session
su had to be run as an unpriveleged user. to run /bin/true as the logging in user, since su doesn't prompt root for the password.
Attempts to use su dates back to commit 976, the initial commit that the waylandloginmanager was brought in. It was wrapped around expect (since su doesn't work with stdio) in commit 1019.
This was one of the oldest and weakest parts with the design of the waylandloginmanager.

This needs libauthen-pam-perl for perl though. This is from https://askubuntu.com/questions/657229

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7657 b52b6941-3400-464c-9f42-43200397181c