waylandloginmanager: Change one small comment, to make another commit with the intent to actually document commit 7705
Somehow the commit message was missed. Commit 7705 added the functionality so that a variable WLM_NO_DBUS_RUN_SESSION can be set to disable launching sessions with dbus-run-session

if this happens, the wsession EnvironmentFile section can just be ammended set=WLM_NO_DBUS_RUN_SESSION=1 so that it won't impact everything if the time does come

It allows systemd user launching of sessions. (Liri supports it, but it had to be disabled because it wasn't starting. Starting without dbus-run-session, and without the liri feature being disabled, it would start)
KDE also supports it, and if not launched with dbus-run-session, plasma-systemmonitor's applications tab actually starts working.

It's not recommended, due to the fact that it doesn't seem to handle the same user logging in twice. Especially with two instances of kde plasma as the same user.
but it is there for in case a session does require systemd launching in the future.


git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7707 b52b6941-3400-464c-9f42-43200397181c