waylandloginmanager: Better handle pre-locking a session with logind
kscreenlocker upon startup actually checks the SetLockedHint value which 'loginctl lock-session' does NOT set. However when it is running it does not appear to check the value for something setting it

KDE sessions when logging in with autologin (when waylandloginmanager is not set by Casper to be in "live cd mode") now locks the session as it starts, which emulates an old KDM feature to combine security with autologins.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7775 b52b6941-3400-464c-9f42-43200397181c