waylandloginmanager:Also switch back to the loginmanagerdisplay if the leader pid of the session terminates, but the logind session for it is still active
the waylandloginmanager only switches back to the logind session after a session termination if it was the current one, so that if one crashes while inactive, and another session is running, the user is not suddenly interrupted being switched back to the loginmanagerdisplay.

On seats with VTs, it's easy, the active VT, stays the same after the active session dies on it, so all that is needed to determine if it is running if if the known leader pid is running or not
On seats without VTs, this was written with the assumption that the logind goes away without the leader pid, so the active session property is blank, which covers most cases.

However it has been proven that logind sessions can remain present as they are terminating from time to time. Account for this, as sometimes the session will not fall back until the slower check runs

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7780 b52b6941-3400-464c-9f42-43200397181c