waylandloginmanager: Prefer defaultwsession-{$SEAT}-:{$USER_LOGNAME} over defaultwsession-{$SEAT}-{$USER_LOGNAME} for saving user/seat prefered session selections
An oversight when the feature was added, as it is technically possible for seat names and user names to collide, as they both can have the - char (seat-something1-something2 and a seat-something1, with a user something2)

Seats or users cannot contain : characters. Even useradd fails when forced, due to /etc/passwd so use : as a separator. Old config lines are still handled for compatibility


git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7785 b52b6941-3400-464c-9f42-43200397181c