waylandloginmanager: Support non-wsession session files
wsession files were meant to be Desktop files but for features the waylandloginmanager was meant to support, namely setting the WAYLAND_DISPLAY 

Now with SimpleDRM, only CompositorExec is needed

This requires a few changes,

Instead of CompositorExec, it is Exec, Exec is now an alias to CompositorExec, and it can be a shell command, not an absolute path, so it's no longer expected to be one that is checked to be a valid executable. (as some might be `env var1=something var2=somethingelse /path/to/binary`)

TryExec is similar to NeedsFile, although it's not always an absolute path, so when it is not, it is ignored

The Type is "Application", not "WSession", so now "Application" and "WSession are both valid.

The .desktop files under wayland-sessions are not executable. User ones will still be checked though

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7798 b52b6941-3400-464c-9f42-43200397181c