vtty: Allow seat-vtty to be created on CONFIG_VT=n
While vttys were first made to replace the kernel mode console with a usermode non-root foot/cage combination, It can also be used as a fallback without VTs

seat-vtty is the fake virtual workaround seat that the actual user sessions started on with the vtty configuration.
This is needed when there were VTs as for the PAM session to start, it needs to be associated with a VT if it's on a seat that has VTs (seat0)
However, the cage display server was already running under the VT, so the session running under the socat "server" can't be associated with that VT. And even if they could, there would be two logind sessions active at the same time on the same seat, which is not possible

However even without VTs, PAM sessions need to start on a seat for a full session (where nmtui works as non-root for example), so this is why the dummy seat-vtty is now being created again
Otherwise if the session was started on the same seat, two sessions still technically need to be active at the same time, the one for the display server, and the one for the user session that runs under the socat server that cage/foot acts as the front end for

The workaround for a logind seat to be created is that a seat is created when a udev device with the master-of-seat tag is associated with it. associate the /dev/tty which always exists, instead of /dev/tty0, which is only for VT systems only

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7872 b52b6941-3400-464c-9f42-43200397181c