waylandloginmanager: Further improve flickerfree startup from Plymouth to Weston.
Since the plymouth deactivate service is being used instead of the plymouth quit service, do not call 'plymouth quit' from waylandloginmanager
If weston starts up too slow on a slower disk or processor, it results in a blank screen between Plymouth and Weston.

DO call 'plymouth deactivate' now, in case the plymouth-deactivate.service file is missing (as it's not part of Plymouth). 
This can be called early in the startup instead of around the time the loginmanagerdisplay starts now, and can be called multiple times. 

This is because when Plymouth deactivates instead of quits, Plymouth gives up control, but leaves the contents of the splash screen up, until the display server draws its first frame

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7884 b52b6941-3400-464c-9f42-43200397181c