waylandloginmanager: Add a wait for udev to initialize all devices, request plymouth to quit on the first UI action.
The udev wait is for when the waylandloginmanager starts too early before udev initializes all devices, and therefore all seats. DetectSeats only runs at most once a minute, so sometimes if udev is too slow to create all seats, it will take 1 minute for the loginmanagerdisplay greeter to start on non-seat0 seats

plymouth deactivate does NOT quit plymouth. That is why plymouth quit was being called by the waylandloginmanager in prior commits. Add a smarter way to call plymouth quit. On the first UI action, from the switch user dialog, to the login dialog, to the leave dialog, to the info dialog, and to starting of the first session (to account for autologin), that is when waylandloginmanager will run plymouth quit

This is to better ensure that plymouth quit is not run too early. DisplayServerWait is not a guarantee as Weston might be working for the wayland-info probe before the desktop shell starts. If Plymouth is quit before the desktop shell starts, the image it is holding on the cards will disappear before weston-desktop-shell shows up, leaving an interval with a blank screen

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@7885 b52b6941-3400-464c-9f42-43200397181c