build_core: Correct some issues with the dependency evaluation.
Don't match *.so.*, as meson names folders library.so/ and all subobjects under the folder were being excluded. Only include items in the exculde list based on matching the name, not the whole path. An additional measure has been added so that if an "a,so" has to be excluded only a.so will be excluded, and not "libnuma.so" as well by adding a leading / to match the start of the file name

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@8019 b52b6941-3400-464c-9f42-43200397181c