Do not allow gnome-session to clear XDG_SEAT and XDG_SESSION_ID, after all
while the waylandloginmanager was adapted for not having these variables, it seems that gnome-shell itself needs these variables if another session running as the same user is running

gnome-shell when started with systemd was failing to start, turns out when another session was active. It was logging "failed to take control" without much usable detail
gnome-shell's user session service file had to be edited to start it with strace. The strace lines, logged to the ~/.cache/wlsessionlogs/log for the session indicated that it was trying to open the 'c2' logind session, which was another session running as the live session user before, and not its own

mutter seems to first use XDG_SESSION_ID variable before it guesses (which is what gnome-shell is based on) By having the XDG_SESSION_ID variable in its environment this should mitigate this

the waylandloginmanager never prevented the same user (usually the live session user) from being logged in multiple times

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@8058 b52b6941-3400-464c-9f42-43200397181c