waylandloginmanager: Add locking to the privleged loginmanager_register FIFO, similar to the main FIFO
The main FIFO has proven that in some cases if two processes write to the same FIFO at once sometimes the FIFO gets muddled. While this is much rarer, as two sessions would have to start pretty quickly on a multiseat configuration, remove the possibility. mknod errors out if it fails to create the lock file if one already exists, if it succeeds, that means that this is the winning process

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@8433 b52b6941-3400-464c-9f42-43200397181c