Remove obsolete updater scripts.
Some of the oldest scripts in the SVN
This was mostly for when testing changes. It was more difficult with VirtualBox, with no clipboard extensions, or serial connections that virt-manager has.
(and rebuilding the ISOs were slower on i3's and dual core Pentiums)

The scripts are not really best, or meant to be run by users, they don't recompile the packages, and was before build_core, so before everything was rolled into separate packages, so the changes made by these scripts now would break hashes.

git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@9099 b52b6941-3400-464c-9f42-43200397181c